backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

sqlalchemy==2.0.25
geoalchemy2==0.14.3
psycopg2-binary==2.9.9
alembic==1.13.1

openai==1.10.0
pydantic==2.5.3
pydantic-settings==2.1.0

python-dotenv==1.0.0
httpx==0.26.0

pytest==7.4.4
pytest-asyncio==0.23.3

mypy==1.8.0
